+++ canvas_sdk/test_utils/factories/__init__.py
@@ -28,6 +28,7 @@ from .eligibility_response import EligibilityRequestFactory, EligibilityResponse
 from .external_event import ExternalEventFactory, ExternalVisitFactory
 from .facility import FacilityFactory
 from .family_history import FamilyHistoryCodingFactory, FamilyHistoryFactory
+from .follow_up import FollowUpFactory
 from .goal import GoalFactory, UpdateGoalFactory
 from .history_present_illness import HistoryOfPresentIllnessFactory
 from .imaging import (
@@ -86,6 +87,7 @@ from .patient_administrative_document import (
     PatientAdministrativeDocumentFactory,
 )
 from .patient_group import PatientGroupFactory, PatientGroupMemberFactory
+from .plan import PlanFactory
 from .plugin_command import PluginCommandFactory
 from .practicelocation import (
     PracticeLocationAddressFactory,
@@ -97,12 +99,15 @@ from .prescription import PrescriptionFactory
 from .procedure import ProcedureCodingFactory, ProcedureFactory
 from .protocol_current import ProtocolCurrentFactory
 from .protocol_override import ProtocolOverrideFactory
+from .reason_for_visit import ReasonForVisitCodingFactory, ReasonForVisitFactory
 from .referral import (
     ReferralFactory,
     ReferralReportCodingFactory,
     ReferralReportFactory,
     ReferralReviewFactory,
 )
+from .remove_allergy_event import RemoveAllergyEventFactory
+from .resolve_condition_event import ResolveConditionEventFactory
 from .service_provider import ServiceProviderFactory
 from .staff import (
     StaffAddressFactory,
@@ -164,6 +169,7 @@ __all__ = (
     "HistoryOfPresentIllnessFactory",
     "FamilyHistoryFactory",
     "FamilyHistoryCodingFactory",
+    "FollowUpFactory",
     "GoalFactory",
     "ImagingOrderFactory",
     "ImagingReportCodingFactory",
@@ -213,6 +219,7 @@ __all__ = (
     "PatientGroupFactory",
     "PatientGroupMemberFactory",
     "PatientPhotoFactory",
+    "PlanFactory",
     "PluginCommandFactory",
     "PracticeLocationFactory",
     "PracticeLocationAddressFactory",
@@ -223,10 +230,14 @@ __all__ = (
     "ProcedureCodingFactory",
     "ProtocolCurrentFactory",
     "ProtocolOverrideFactory",
+    "ReasonForVisitCodingFactory",
+    "ReasonForVisitFactory",
     "ReferralFactory",
     "ReferralReportCodingFactory",
     "ReferralReportFactory",
     "ReferralReviewFactory",
+    "RemoveAllergyEventFactory",
+    "ResolveConditionEventFactory",
     "ServiceProviderFactory",
     "StaffFactory",
     "StaffPhotoFactory",