+++ canvas_sdk/v1/data/__init__.py
@@ -72,6 +72,7 @@ from .encounter import Encounter
 from .external_event import ExternalEvent, ExternalVisit
 from .facility import Facility
 from .family_history import FamilyHistory, FamilyHistoryCoding
+from .follow_up import FollowUp
 from .goal import Goal, UpdateGoal
 from .history_present_illness import HistoryOfPresentIllness
 from .imaging import (
@@ -166,6 +167,7 @@ from .patient_consent import (
 from .patient_group import PatientGroup, PatientGroupMember
 from .payment_collection import PaymentCollection
 from .payor_specific_charge import PayorSpecificCharge
+from .plan import Plan
 from .plugin_command import PluginCommand
 from .posting import (
     BasePosting,
@@ -194,9 +196,11 @@ from .questionnaire import (
     ResponseOption,
     ResponseOptionSet,
 )
-from .reason_for_visit import ReasonForVisitSettingCoding
+from .reason_for_visit import ReasonForVisit, ReasonForVisitCoding, ReasonForVisitSettingCoding
 from .referral import Referral, ReferralReport, ReferralReportCoding, ReferralReview
 from .refill_request import RefillRequest, RefillRequestCoding
+from .remove_allergy_event import RemoveAllergyEvent
+from .resolve_condition_event import ResolveConditionEvent
 from .service_provider import ServiceProvider
 from .snapshot import Snapshot, SnapshotImage
 from .specialty_report_template import (
@@ -308,6 +312,7 @@ __all__ = __exports__ = (
     "FamilyHistoryCoding",
     "FieldValueTooLarge",
     "FieldType",
+    "FollowUp",
     "Goal",
     "HistoryOfPresentIllness",
     "ImagingOrder",
@@ -404,6 +409,7 @@ __all__ = __exports__ = (
     "PatientGroupMember",
     "PayorSpecificCharge",
     "PaymentCollection",
+    "Plan",
     "PluginCommand",
     "PracticeLocation",
     "PracticeLocationAddress",
@@ -420,6 +426,8 @@ __all__ = __exports__ = (
     "Question",
     "Questionnaire",
     "QuestionnaireQuestionMap",
+    "ReasonForVisit",
+    "ReasonForVisitCoding",
     "ReasonForVisitSettingCoding",
     "Referral",
     "ReferralReport",
@@ -427,6 +435,8 @@ __all__ = __exports__ = (
     "ReferralReview",
     "RefillRequest",
     "RefillRequestCoding",
+    "RemoveAllergyEvent",
+    "ResolveConditionEvent",
     "ResponseOption",
     "ResponseOptionSet",
     "ServiceProvider",